modify coupon status

+++ equity/src/main/java/com/fengchao/equity/service/impl/CouponServiceImpl.java
@@ -74,23 +74,31 @@ public class CouponServiceImpl implements CouponService {
     public int updateCoupon(CouponBean bean) {
         CouponX coupon = beanToCoupon(bean);
         coupon.setId(bean.getId());
+        CouponX couponById = new CouponX();
         if(bean.getStatus() != null && bean.getStatus() == 2){
             coupon.setStatus(3);
-            CouponX couponById = mapper.selectByPrimaryKey(bean.getId());
+            couponById = mapper.selectByPrimaryKey(bean.getId());
             if(couponById == null){
                 return 0;
             }
-            JobClientUtils.couponEffectiveTrigger(jobClient, coupon.getId(), couponById.getReleaseStartDate());
-            JobClientUtils.couponEndTrigger(jobClient, coupon.getId(), couponById.getReleaseEndDate());
         }
 
         int num = mapper.updateByPrimaryKeySelective(coupon);
+
+        if(bean.getStatus() != null && bean.getStatus() == 2 && num == 1){
+            JobClientUtils.couponEffectiveTrigger(jobClient, coupon.getId(), couponById.getReleaseStartDate());
+            JobClientUtils.couponEndTrigger(jobClient, coupon.getId(), couponById.getReleaseEndDate());
+        }
 
         return num;
     }
 
     @Override
     public int deleteCoupon(Integer id) {
+        CouponX couponX = mapper.selectByPrimaryKey(id);
+        if(couponX.getStatus() != 1){
+            return 2;
+        }
         return mapper.deleteByPrimaryKey(id);
     }
 
@@ -188,7 +196,7 @@ public class CouponServiceImpl implements CouponService {
         useInfo.setId(bean.getId());
         useInfo.setConsumedTime(new Date());
         useInfo.setUserCouponCode(bean.getUserCouponCode());
-        useInfo.setStatus(2);
+        useInfo.setStatus(3);
         useInfoMapper.updateStatusByUserCode(useInfo);
         CouponX coupon = mapper.selectByPrimaryKey(couponUseInfo.getCouponId());
         return coupon;
@@ -392,6 +400,11 @@ public class CouponServiceImpl implements CouponService {
         log.info("根据id集合查询coupon列表 返回List<CouponBean>:{}", JSONUtil.toJsonString(couponBeanList));
 
         return couponBeanList;
+    }
+
+    @Override
+    public int invalid(int couponId) {
+        return useInfoMapper.updateStatusByCouponId(couponId);
     }
 
     //============================== private =========================
